Refactor data center ETL into medallion stage modules
