+++ etl-pipelines/exotic/data-centers/src/data_center_etl_pipeline/__init__.py
@@ -1,0 +1,5 @@
+"""Data-center ETL pipeline package with bronze/silver/gold stages."""
+
+from .pipeline import run_etl
+
+__all__ = ["run_etl"]
