+++ etl-pipelines/exotic/data-centers/src/data_center_etl_pipeline/constants.py
@@ -1,0 +1,15 @@
+NUMERIC_FIELDS = [
+    "gross_revenue_eur",
+    "energy_cost_eur",
+    "opex_eur",
+    "debt_service_eur",
+    "market_value_eur",
+    "outstanding_debt_eur",
+    "it_load_mw",
+    "leased_capacity_mw",
+]
+
+DSCR_WATCH = 1.35
+DSCR_BREACH = 1.20
+LTV_WATCH = 74.0
+LTV_BREACH = 78.0
